Split testbed app tests into desktop and mobile modules.
