+++ android/src/toga_android/app.py
@@ -245,10 +245,8 @@ class App:
         pass  # pragma: no cover
 
     def set_main_window(self, window):
-        if window is None:
-            raise ValueError("Session-based apps are not supported on Android")
-        elif window == toga.App.BACKGROUND:
-            raise ValueError("Background apps are not supported on Android")
+        if window is None or window == toga.App.BACKGROUND:
+            raise ValueError("Apps without main windows are not supported on Android")
         else:
             # The default layout of an Android app includes a titlebar; a simple App
             # then hides that titlebar. We know what type of app we have when the main
